Restructured crate. Constants moved to their own file. Traits moved to their own file. Types moves to their own file. ConvertApi added, but untested. Additional optimizations for building queries.
