+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "cbadv"
 license = "MIT"
-version = "1.3.1"
+version = "1.3.2"
 edition = "2021"
 description = "Asynchronous Coinbase Advanced REST and WebSocket API"
 readme = "README.md"
@@ -50,6 +50,11 @@ name = "custom_config"
 path = "examples/custom_config.rs"
 required-features = ["config"]
 
+[profile.release]
+strip = "symbols"
+lto = "fat"
+codegen-units = 1
+opt-level = 3
 
 [dependencies]
 reqwest = { version = "0.11", features = ["json"] } # Making HTTP requests.
