+++ src/apis/account.rs
@@ -3,91 +3,14 @@
 //! `account` gives access to the Account API and the various endpoints associated with it.
 //! This allows you to obtain account information either by account UUID or in bulk (all accounts).
 
-use crate::signer::Signer;
-use crate::utils::{deserialize_numeric, CbAdvError, CbResult, QueryBuilder};
 use async_recursion::async_recursion;
-use serde::{Deserialize, Serialize};
-use std::fmt;
 
-/// Represents a Balance for either Available or Held funds.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Balance {
-    /// Value for the currency available or held.
-    #[serde(deserialize_with = "deserialize_numeric")]
-    pub value: f64,
-    /// Denomination of the currency.
-    pub currency: String,
-}
-
-/// Represents an Account received from the API.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Account {
-    /// Unique identifier for account.
-    pub uuid: String,
-    /// Name for the account.
-    pub name: String,
-    /// Currency symbol for the account.
-    pub currency: String,
-    /// Current available balance for the account.
-    pub available_balance: Balance,
-    /// Whether or not this account is the user's primary account.
-    pub default: bool,
-    /// Whether or not this account is active and okay to use.
-    pub active: bool,
-    /// Time at which this account was created.
-    pub created_at: String,
-    /// Time at which this account was updated.
-    pub updated_at: String,
-    /// Time at which this account was deleted.
-    pub deleted_at: Option<String>,
-    /// Possible values: [ACCOUNT_TYPE_UNSPECIFIED, ACCOUNT_TYPE_CRYPTO, ACCOUNT_TYPE_FIAT, ACCOUNT_TYPE_VAULT]
-    pub r#type: String,
-    /// Whether or not this account is ready to trade.
-    pub ready: bool,
-    /// Current balance on hold.
-    pub hold: Balance,
-}
-
-/// Represents a list of accounts received from the API.
-#[derive(Deserialize, Debug)]
-pub struct ListedAccounts {
-    /// Accounts returned from the API.
-    pub accounts: Vec<Account>,
-    /// Whether there are additional pages for this query.
-    pub has_next: bool,
-    /// Cursor for paginating. Users can use this string to pass in the next call to this endpoint, and repeat this process to fetch all accounts through pagination.
-    pub cursor: String,
-    /// Number of accounts returned.
-    pub size: u32,
-}
-
-/// Represents an account response from the API.
-#[derive(Deserialize, Debug)]
-struct AccountResponse {
-    /// Account returned from the API.
-    pub account: Account,
-}
-
-/// Represents parameters that are optional for List Account API request.
-#[derive(Serialize, Default, Debug, Clone)]
-pub struct ListAccountsQuery {
-    /// Amount to obtain, default 49 maximum is 250.
-    pub limit: Option<u32>,
-    /// Returns accounts after the cursor provided.
-    pub cursor: Option<String>,
-}
-
-impl fmt::Display for ListAccountsQuery {
-    /// Converts the object into HTTP request parameters.
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut query = QueryBuilder::new();
-        query
-            .push_u32_optional("limit", self.limit)
-            .push_optional("cursor", &self.cursor);
-
-        write!(f, "{}", query.build())
-    }
-}
+use crate::account::{Account, AccountResponse, ListAccountsQuery, ListedAccounts};
+use crate::constants::accounts::RESOURCE_ENDPOINT;
+use crate::errors::CbAdvError;
+use crate::signer::Signer;
+use crate::traits::NoQuery;
+use crate::types::CbResult;
 
 /// Provides access to the Account API for the service.
 pub struct AccountApi {
@@ -96,9 +19,6 @@ pub struct AccountApi {
 }
 
 impl AccountApi {
-    /// Resource for the API.
-    const RESOURCE: &'static str = "/api/v3/brokerage/accounts";
-
     /// Creates a new instance of the Account API. This grants access to account information.
     ///
     /// # Arguments
@@ -123,8 +43,8 @@ impl AccountApi {
     ///
     /// <https://docs.cloud.coinbase.com/advanced-trade-api/reference/retailbrokerageapi_getaccount>
     pub async fn get(&mut self, account_uuid: &str) -> CbResult<Account> {
-        let resource = format!("{}/{}", Self::RESOURCE, account_uuid);
-        match self.signer.get(&resource, "").await {
+        let resource = format!("{}/{}", RESOURCE_ENDPOINT, account_uuid);
+        match self.signer.get(&resource, &NoQuery).await {
             Ok(value) => match value.json::<AccountResponse>().await {
                 Ok(resp) => Ok(resp.account),
                 Err(_) => Err(CbAdvError::BadParse("account object".to_string())),
@@ -218,7 +138,7 @@ impl AccountApi {
     ///
     /// <https://docs.cloud.coinbase.com/advanced-trade-api/reference/retailbrokerageapi_getaccounts>
     pub async fn get_bulk(&mut self, query: &ListAccountsQuery) -> CbResult<ListedAccounts> {
-        match self.signer.get(Self::RESOURCE, &query.to_string()).await {
+        match self.signer.get(RESOURCE_ENDPOINT, query).await {
             Ok(value) => match value.json::<ListedAccounts>().await {
                 Ok(resp) => Ok(resp),
                 Err(_) => Err(CbAdvError::BadParse("accounts vector".to_string())),
