+++ src/apis/mod.rs
@@ -1,0 +1,11 @@
+mod account;
+mod convert;
+mod fee;
+mod order;
+mod product;
+
+pub(crate) use account::AccountApi;
+pub(crate) use convert::ConvertApi;
+pub(crate) use fee::FeeApi;
+pub(crate) use order::OrderApi;
+pub(crate) use product::ProductApi;
