+++ src/models/fee.rs
@@ -3,10 +3,10 @@
 //! `fee` gives access to the Fee API and the various endpoints associated with it.
 //! Currently the only endpoint available is the Transaction Summary endpoint.
 
-use crate::signer::Signer;
-use crate::utils::{deserialize_numeric, CbAdvError, CbResult, QueryBuilder};
 use serde::{Deserialize, Serialize};
-use std::fmt;
+
+use crate::traits::Query;
+use crate::utils::{deserialize_numeric, QueryBuilder};
 
 /// Pricing tier for user, determined by notional (USD) volume.
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -35,7 +35,7 @@ pub struct MarginRate {
     pub value: f64,
 }
 
-// Represents a Tax.
+/// Represents a tax amount.
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Tax {
     /// Amount of tax.
@@ -81,60 +81,14 @@ pub struct TransactionSummaryQuery {
     pub product_type: Option<String>,
 }
 
-impl fmt::Display for TransactionSummaryQuery {
+impl Query for TransactionSummaryQuery {
     /// Converts the object into HTTP request parameters.
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut query = QueryBuilder::new();
-        query
+    fn to_query(&self) -> String {
+        QueryBuilder::new()
             .push_optional("start_date", &self.start_date)
             .push_optional("end_date", &self.end_date)
             .push_optional("user_native_currency", &self.user_native_currency)
-            .push_optional("product_type", &self.product_type);
-
-        write!(f, "{}", query.build())
-    }
-}
-
-/// Provides access to the Fee API for the service.
-pub struct FeeApi {
-    /// Object used to sign requests made to the API.
-    signer: Signer,
-}
-
-impl FeeApi {
-    /// Resource for the API.
-    const RESOURCE: &'static str = "/api/v3/brokerage/transaction_summary";
-
-    /// Creates a new instance of the Fee API. This grants access to product information.
-    ///
-    /// # Arguments
-    ///
-    /// * `signer` - A Signer that include the API Key & Secret along with a client to make
-    /// requests.
-    pub(crate) fn new(signer: Signer) -> Self {
-        Self { signer }
-    }
-
-    /// Obtains fee transaction summary from the API.
-    ///
-    /// # Arguments
-    ///
-    /// * `query` - Optional paramaters used to modify the resulting scope of the
-    /// summary.
-    ///
-    /// # Endpoint / Reference
-    ///
-    #[allow(rustdoc::bare_urls)]
-    /// https://api.coinbase.com/api/v3/brokerage/transaction_summary
-    ///
-    /// <https://docs.cloud.coinbase.com/advanced-trade-api/reference/retailbrokerageapi_gettransactionsummary>
-    pub async fn get(&mut self, query: &TransactionSummaryQuery) -> CbResult<TransactionSummary> {
-        match self.signer.get(Self::RESOURCE, &query.to_string()).await {
-            Ok(value) => match value.json::<TransactionSummary>().await {
-                Ok(resp) => Ok(resp),
-                Err(_) => Err(CbAdvError::BadParse("fee summary object".to_string())),
-            },
-            Err(error) => Err(error),
-        }
+            .push_optional("product_type", &self.product_type)
+            .build()
     }
 }
