+++ src/models/mod.rs
@@ -1,0 +1,6 @@
+pub mod account;
+pub mod convert;
+pub mod fee;
+pub mod order;
+pub mod product;
+pub mod websocket;
