+++ src/rest.rs
@@ -4,10 +4,7 @@
 //! This is the primary method of accessing the endpoints and handles all of the configurations and
 //! negotiations for the user.
 
-use crate::account::AccountApi;
-use crate::fee::FeeApi;
-use crate::order::OrderApi;
-use crate::product::ProductApi;
+use crate::apis::{AccountApi, ConvertApi, FeeApi, OrderApi, ProductApi};
 use crate::signer::Signer;
 
 #[cfg(feature = "config")]
@@ -23,6 +20,8 @@ pub struct RestClient {
     pub fee: FeeApi,
     /// Gives access to the Order API.
     pub order: OrderApi,
+    /// Gives access to the Convert API.
+    pub convert: ConvertApi,
 }
 
 impl RestClient {
@@ -38,6 +37,7 @@ impl RestClient {
             product: ProductApi::new(Signer::new(key.to_string(), secret.to_string(), true)),
             fee: FeeApi::new(Signer::new(key.to_string(), secret.to_string(), true)),
             order: OrderApi::new(Signer::new(key.to_string(), secret.to_string(), true)),
+            convert: ConvertApi::new(Signer::new(key.to_string(), secret.to_string(), true)),
         }
     }
 
@@ -53,27 +53,4 @@ impl RestClient {
     {
         Self::new(&config.coinbase().api_key, &config.coinbase().api_secret)
     }
-}
-
-/// Creates a new instance of a Client. This is a wrapper for the various APIs and Signer.
-///
-/// # Arguments
-///
-/// * `key` - A string that holds the key for the API service.
-/// * `secret` - A string that holds the secret for the API service.
-pub fn new(key: &str, secret: &str) -> RestClient {
-    RestClient::new(key, secret)
-}
-
-/// Creates a new instance of a Client using a configuration file. This is a wrapper for the various APIs and Signer.
-///
-/// # Arguments
-///
-/// * `config` - Configuration that implements ConfigFile trait.
-#[cfg(feature = "config")]
-pub fn from_config<T>(config: &T) -> RestClient
-where
-    T: ConfigFile,
-{
-    RestClient::from_config(config)
 }
