+++ src/websocket.rs
@@ -4,288 +4,25 @@
 //! Many parts of the REST API suggest using websockets instead due to ratelimits and being quicker
 //! for large amount of constantly changing data.
 
-use crate::order::OrderUpdate;
-use crate::product::{Candle, CandleUpdate, MarketTradesUpdate, ProductUpdate, TickerUpdate};
-use crate::signer::Signer;
-use crate::task_tracker::TaskTracker;
-use crate::time;
-use crate::utils::{deserialize_numeric, CbAdvError, CbResult};
-use futures_util::stream::{SplitSink, SplitStream};
+use futures_util::stream::SplitSink;
 use futures_util::{SinkExt, StreamExt};
-use serde::{Deserialize, Serialize};
-use std::fmt;
 use tokio::net::TcpStream;
 use tokio::task::JoinHandle;
 use tokio_tungstenite::{connect_async, tungstenite, MaybeTlsStream, WebSocketStream};
+
+use crate::constants::websocket::RESOURCE_ENDPOINT;
+use crate::errors::CbAdvError;
+use crate::models::websocket::{Channel, Subscription};
+use crate::signer::Signer;
+use crate::task_tracker::TaskTracker;
+use crate::time;
+use crate::traits::{CandleCallback, MessageCallback};
+use crate::types::{CbResult, MessageCallbackFn, WebSocketReader};
 
 #[cfg(feature = "config")]
 use crate::config::ConfigFile;
 
 type Socket = WebSocketStream<MaybeTlsStream<TcpStream>>;
-pub type Callback = fn(CbResult<Message>);
-pub type WebSocketReader = SplitStream<Socket>;
-
-/// Used to pass to a callback to the candle watcher on a successful ejection.
-pub trait CandleCallback {
-    /// Called when a candle is succesfully ejected.
-    ///
-    /// # Arguments
-    ///
-    /// * `current_start` - Current UTC timestamp for a start.
-    /// * `product_id` - Product the candle belongs to.
-    /// * `candle` - Candle that was recently completed.
-    fn candle_callback(&mut self, current_start: u64, product_id: String, candle: Candle);
-}
-
-/// Used to pass objects to the listener for greater control over message processing.
-pub trait MessageCallback {
-    /// This is called when processing a message from the WebSocket.
-    ///
-    /// # Arguments
-    ///
-    /// * `msg` - Message or Error received from the WebSocket.
-    fn message_callback(&mut self, msg: CbResult<Message>);
-}
-
-/// WebSocket Channels that can be subscribed to.
-#[allow(non_camel_case_types)]
-#[derive(Serialize, Deserialize, Debug)]
-pub enum Channel {
-    /// Sends all products and currencies on a preset interval.
-    STATUS,
-    /// Updates every second. Candles are grouped into buckets (granularities) of five minutes.
-    CANDLES,
-    /// Real-time price updates every time a match happens.
-    TICKER,
-    /// Real-time price updates every 5000 milli-seconds.
-    TICKER_BATCH,
-    /// All updates and easiest way to keep order book snapshot
-    LEVEL2,
-    /// Only sends messages that include the authenticated user.
-    USER,
-    /// Real-time updates every time a market trade happens.
-    MARKET_TRADES,
-    /// Real-time pings from server to keep connections open.
-    HEARTBEATS,
-}
-
-impl fmt::Display for Channel {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Channel::STATUS => write!(f, "status"),
-            Channel::CANDLES => write!(f, "candles"),
-            Channel::TICKER => write!(f, "ticker"),
-            Channel::TICKER_BATCH => write!(f, "ticker_batch"),
-            Channel::LEVEL2 => write!(f, "level2"),
-            Channel::USER => write!(f, "user"),
-            Channel::MARKET_TRADES => write!(f, "market_trades"),
-            Channel::HEARTBEATS => write!(f, "heartbeats"),
-        }
-    }
-}
-
-/// Messages that could be received from the WebSocket.
-#[derive(Deserialize, Debug)]
-#[serde(untagged)]
-pub enum Message {
-    /// Sends all products and currencies on a preset interval.
-    Status(StatusMessage),
-    /// Updates every second. Candles are grouped into buckets (granularities) of five minutes.
-    Candles(CandlesMessage),
-    /// Real-time price updates every time a match happens.
-    Ticker(TickerMessage),
-    /// All updates and easiest way to keep order book snapshot
-    TickerBatch(TickerMessage),
-    /// All updates and easiest way to keep order book snapshot
-    Level2(Level2Message),
-    /// Only sends messages that include the authenticated user.
-    User(UserMessage),
-    /// Real-time updates every time a market trade happens.
-    MarketTrades(MarketTradesMessage),
-    /// Real-time pings from server to keep connections open.
-    Heartbeats(HeartbeatsMessage),
-    /// Subscription updates.
-    Subscribe(SubscribeMessage),
-}
-
-/// Data received from the WebSocket for Level2 Events.
-#[derive(Deserialize, Debug)]
-pub struct Level2Update {
-    pub side: String,
-    pub event_time: String,
-    #[serde(deserialize_with = "deserialize_numeric")]
-    pub price_level: f64,
-    #[serde(deserialize_with = "deserialize_numeric")]
-    pub new_quantity: f64,
-}
-
-/// Data received from the WebSocket for Subscription Update Events.
-#[derive(Deserialize, Debug, Default)]
-pub struct SubscribeUpdate {
-    #[serde(default)]
-    pub status: Vec<String>,
-    #[serde(default)]
-    pub ticker: Vec<String>,
-    #[serde(default)]
-    pub ticker_batch: Vec<String>,
-    #[serde(default)]
-    pub level2: Option<Vec<String>>,
-    #[serde(default)]
-    pub user: Option<Vec<String>>,
-    #[serde(default)]
-    pub market_trades: Option<Vec<String>>,
-    #[serde(default)]
-    pub heartbeats: Option<Vec<String>>,
-}
-
-/// Status Event received from the WebSocket, contained inside the Status Message.
-#[derive(Deserialize, Debug)]
-pub struct StatusEvent {
-    pub r#type: String,
-    pub products: Vec<ProductUpdate>,
-}
-
-/// Candles Event received from the WebSocket, contained inside the Candles Message.
-#[derive(Deserialize, Debug)]
-pub struct CandlesEvent {
-    pub r#type: String,
-    pub candles: Vec<CandleUpdate>,
-}
-
-/// Ticker Event received from the WebSocket, contained inside the Ticker Message.
-#[derive(Deserialize, Debug)]
-pub struct TickerEvent {
-    pub r#type: String,
-    pub tickers: Vec<TickerUpdate>,
-}
-
-/// Level2 Event received from the WebSocket, contained inside the Level2 Message.
-#[derive(Deserialize, Debug)]
-pub struct Level2Event {
-    pub r#type: String,
-    pub product_id: String,
-    pub updates: Vec<Level2Update>,
-}
-
-/// User Event received from the WebSocket, contained inside the User Message.
-#[derive(Deserialize, Debug)]
-pub struct UserEvent {
-    pub r#type: String,
-    pub orders: Vec<OrderUpdate>,
-}
-
-/// Market Trades Event received from the WebSocket, contained inside the Market Trades Message.
-#[derive(Deserialize, Debug)]
-pub struct MarketTradesEvent {
-    pub r#type: String,
-    pub trades: Vec<MarketTradesUpdate>,
-}
-
-/// Heartbeats Event received from the WebSocket, contained inside the Heartbeats Message.
-#[derive(Deserialize, Debug)]
-pub struct HeartbeatsEvent {
-    pub current_time: String,
-    pub heartbeat_counter: u64,
-}
-
-/// Subscribe Event received from the WebSocket, contained inside the Subscribe Message.
-#[derive(Deserialize, Debug)]
-pub struct SubscribeEvent {
-    pub subscriptions: SubscribeUpdate,
-}
-
-/// Message received from the WebSocket API. Contains updates on product statuses.
-#[derive(Deserialize, Debug)]
-pub struct StatusMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<StatusEvent>,
-}
-
-/// Message received from the WebSocket API. Contains updates on candles.
-#[derive(Deserialize, Debug)]
-pub struct CandlesMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<CandlesEvent>,
-}
-
-/// Message received from the WebSocket API. Contains updates on products and currencies.
-#[derive(Deserialize, Debug)]
-pub struct TickerMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<TickerEvent>,
-}
-
-/// Message received from the WebSocket API. All order updates for a products. Best way to
-/// keep a snapshot of the order book.
-#[derive(Deserialize, Debug)]
-pub struct Level2Message {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<Level2Event>,
-}
-
-/// Message received from the WebSocket API. Contains order updates strictly for the user.
-#[derive(Deserialize, Debug)]
-pub struct UserMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<UserEvent>,
-}
-
-/// Message received from the WebSocket API. Real-time updates everytime a market trade happens.
-#[derive(Deserialize, Debug)]
-pub struct MarketTradesMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<MarketTradesEvent>,
-}
-
-/// Message received from the WebSocket API. Real-time pings from the server to keep connections
-/// open.
-#[derive(Deserialize, Debug)]
-pub struct HeartbeatsMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<HeartbeatsEvent>,
-}
-
-/// Message received from the WebSocket API. Provides updates for the current subscriptions.
-#[derive(Deserialize, Debug)]
-pub struct SubscribeMessage {
-    pub channel: String,
-    pub client_id: String,
-    pub timestamp: String,
-    pub sequence_num: u64,
-    pub events: Vec<SubscribeEvent>,
-}
-
-/// Subscription is sent to the WebSocket to enable updates for specified Channels.
-#[derive(Serialize, Debug)]
-struct Subscription {
-    pub r#type: String,
-    pub product_ids: Vec<String>,
-    pub channel: String,
-    pub api_key: String,
-    pub timestamp: String,
-    pub signature: String,
-}
 
 /// Represents a Client for the Websocket API. Provides easy-access to subscribing and listening to
 /// the WebSocket.
@@ -297,9 +34,6 @@ pub struct WebSocketClient {
 }
 
 impl WebSocketClient {
-    /// Resource for the API.
-    const RESOURCE: &'static str = "wss://advanced-trade-ws.coinbase.com";
-
     /// Creates a new instance of a Client. This is a wrapper for Signer and contains a socket to
     /// the WebSocket.
     ///
@@ -331,7 +65,7 @@ impl WebSocketClient {
     /// Connects to the WebSocket. This is required before subscribing, unsubscribing, and
     /// listening for updates. A reader is returned to allow for `listener` to parse events.
     pub async fn connect(&mut self) -> CbResult<WebSocketReader> {
-        match connect_async(Self::RESOURCE).await {
+        match connect_async(RESOURCE_ENDPOINT).await {
             Ok((socket, _)) => {
                 let (sink, stream) = socket.split();
                 self.socket_tx = Some(sink);
@@ -353,7 +87,7 @@ impl WebSocketClient {
     /// * `reader` - Allows the listener to receive messages. `Obtained from connect``.
     /// * `callback` - A callback function that is trigger and passed the Message received via
     /// WebSocket, if an error occurred.
-    pub async fn listener(reader: WebSocketReader, callback: Callback) {
+    pub async fn listener(reader: WebSocketReader, callback: MessageCallbackFn) {
         // Read messages and send to the callback as they come in.
         let read_future = reader.for_each(|message| {
             let data: String = match message {
@@ -540,79 +274,17 @@ impl WebSocketClient {
         let listener = tokio::spawn(TaskTracker::start(reader, watcher));
 
         // Keep the connection open.
-        match self.sub(Channel::HEARTBEATS, &[]).await {
+        match self.sub(Channel::Heartbeats, &[]).await {
             Ok(_) => (),
             Err(err) => return Err(err),
         };
 
         // Subscribe to the candle updates for the products.
-        match self.sub(Channel::CANDLES, products).await {
+        match self.sub(Channel::Candles, products).await {
             Ok(_) => (),
             Err(err) => return Err(err),
         };
 
         Ok(listener)
     }
-}
-
-/// Creates a new instance of a Client using a configuration file. This is a wrapper for
-/// Signer and contains a socket to the WebSocket.
-///
-/// # Arguments
-///
-/// * `config` - Configuration that implements ConfigFile trait.
-#[cfg(feature = "config")]
-pub fn from_config<T>(config: &T) -> WebSocketClient
-where
-    T: ConfigFile,
-{
-    WebSocketClient::new(&config.coinbase().api_key, &config.coinbase().api_secret)
-}
-
-/// Starts the listener which returns Messages via a callback function provided by the user.
-/// This allows the user to get objects out of the WebSocket stream for additional processing.
-/// the WebSocket. If it is unable to parse an object received, the user is supplied
-/// CBAdvError::BadParse along with the data it failed to parse.
-///
-/// # Arguments
-///
-/// * `reader` - Allows the listener to receive messages. `Obtained from connect``.
-/// * `callback` - A callback function that is trigger and passed the Message received via
-/// WebSocket, if an error occurred.
-pub async fn listener(reader: WebSocketReader, callback: Callback) {
-    WebSocketClient::listener(reader, callback).await
-}
-
-/// Starts the listener with a callback object that implements the `MessageCallback` trait.
-/// This allows the user to get objects out of the WebSocket stream for additional processing.
-/// the WebSocket. If it is unable to parse an object received, the user is supplied
-/// CBAdvError::BadParse along with the data it failed to parse.
-///
-/// # Arguments
-///
-/// * `reader` - Allows the listener to receive messages. `Obtained from connect``.
-/// * `callback_obj` - A callback object that implements `MessageCallback` trait.
-pub async fn listener_with<T>(reader: WebSocketReader, callback_obj: T)
-where
-    T: MessageCallback,
-{
-    WebSocketClient::listener_with(reader, callback_obj).await
-}
-
-/// Watches candles for a set of products, producing candles once they are considered complete.
-///
-/// # Argument
-///
-/// * `client` - WebSocket client used to obtain a `reader` to listen to the socket.
-/// * `products` - Products to watch for candles for.
-/// * `watcher` - User-defined struct that implements `CandleCallback` to send completed candles to.
-pub async fn watch_candles<T>(
-    client: &mut WebSocketClient,
-    products: &[String],
-    watcher: T,
-) -> CbResult<JoinHandle<()>>
-where
-    T: CandleCallback + Send + 'static,
-{
-    client.watch_candles(products, watcher).await
 }
